% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threat_analysis.R
\name{calculate_flush_threat}
\alias{calculate_flush_threat}
\title{Calculate flush threat probability}
\usage{
calculate_flush_threat(suited_count, opponents)
}
\arguments{
\item{suited_count}{Number of suited cards on board}

\item{opponents}{Number of opponents}
}
\value{
Numeric probability of flush threat
}
\description{
Calculate flush threat probability
}
