% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threat_analysis.R
\name{calculate_overpair_threat}
\alias{calculate_overpair_threat}
\title{Calculate overpair threat probability}
\usage{
calculate_overpair_threat(max_board_rank, opponents)
}
\arguments{
\item{max_board_rank}{Highest rank on board (0-12, where 12 = Ace)}

\item{opponents}{Number of opponents}
}
\value{
Numeric probability of overpair threat
}
\description{
Calculate overpair threat probability
}
