% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threat_analysis.R
\name{calculate_set_threat}
\alias{calculate_set_threat}
\title{Calculate set/full house threat probability}
\usage{
calculate_set_threat(pair_count, opponents)
}
\arguments{
\item{pair_count}{Number of cards of same rank on board}

\item{opponents}{Number of opponents}
}
\value{
Numeric probability of set/full house threat
}
\description{
Calculate set/full house threat probability
}
