% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threat_analysis.R
\name{calculate_straight_threat}
\alias{calculate_straight_threat}
\title{Calculate straight threat probability}
\usage{
calculate_straight_threat(ranks, opponents)
}
\arguments{
\item{ranks}{Vector of card ranks on board}

\item{opponents}{Number of opponents}
}
\value{
Numeric probability of straight threat
}
\description{
Calculate straight threat probability
}
