% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_strength.R
\name{compare_ranks}
\alias{compare_ranks}
\title{Compare two rank vectors for poker hand tie-breaking}
\usage{
compare_ranks(ranks1, ranks2)
}
\arguments{
\item{ranks1}{First rank vector (numeric)}

\item{ranks2}{Second rank vector (numeric)}
}
\value{
Integer: 1 if ranks1 > ranks2, -1 if ranks1 < ranks2, 0 if equal
}
\description{
Compares two rank vectors to determine which represents a stronger hand.
Used for tie-breaking when hands have the same score.
}
\examples{
compare_ranks(c(12, 11, 10), c(12, 11, 9))  # Returns 1 (first is higher)
compare_ranks(c(12, 11, 9), c(12, 11, 10))  # Returns -1 (second is higher)
compare_ranks(c(12, 11, 10), c(12, 11, 10)) # Returns 0 (equal)
}
