% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_strength.R
\name{evaluate_hand_strength}
\alias{evaluate_hand_strength}
\title{Evaluate hand strength using compact algorithm (adapted from Python)}
\usage{
evaluate_hand_strength(my_hand, board = "", opponents = 1)
}
\arguments{
\item{my_hand}{String like "AS KH" (my hole cards)}

\item{board}{String like "9H TC JC QS KC" (community cards)}

\item{opponents}{Number of opponents (default: 1)}
}
\value{
List with hand_score, hand_type, win_estimate, and description
}
\description{
Evaluate hand strength using compact algorithm (adapted from Python)
}
\examples{
evaluate_hand_strength("AS AH", "9H TC JC")
evaluate_hand_strength("QH QS", "9C TD JH AS KC", 2)
}
