% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threat_analysis.R
\name{find_threat_hands}
\alias{find_threat_hands}
\title{Find potential threat hands based on board texture}
\usage{
find_threat_hands(my_hand, board = "", opponents = 1)
}
\arguments{
\item{my_hand}{String like "AS KH" (my hole cards)}

\item{board}{String like "9H TC JC" (community cards)}

\item{opponents}{Number of opponents (default: 1)}
}
\value{
Data frame with threat types and probabilities
}
\description{
Find potential threat hands based on board texture
}
\examples{
find_threat_hands("AS KH", "QS JS 7S")
find_threat_hands("QH QS", "9C TD JH", 2)
}
